Renamed pcode to ghidra
