+++ smallworld/emulators/ghidra/ghidra.py
