+++ smallworld/emulators/ghidra/machdefs/__init__.py
