+++ smallworld/emulators/ghidra/machdefs/machdef.py
