+++ smallworld/emulators/ghidra/machdefs/mips.py
