+++ smallworld/emulators/ghidra/machdefs/mips64.py
