+++ smallworld/emulators/ghidra/machdefs/ppc.py
