+++ smallworld/emulators/ghidra/machdefs/riscv.py
