+++ smallworld/emulators/ghidra/typing.py
